sql: don't set SystemConfigTrigger on table lease modifications
Table leases are always accessed consistently, so we do not need to
gossip the SystemConfig span when they change. This reduces the number
of times we need to scan the `SystemConfig` span during schema changes,
including DROP TABLE statements. We saw that scanning this span was
expensive when the number of descriptors became large, so this is a
performance win.

Release note (performance improvement): Improve the performance of DROP
TABLE.