storage: kick slow closed timestamps being used by rangefeed
Changefeeds and, more proximately, RangeFeeds need closed timestamps to
keep advancing to make progress, but sometimes we have bugs that lead to
closed timestamps getting stuck. We should and will fix the bugs, but
given that even one stuck closed timestamp will stall changefeed-wide
progress, this commit introduces a fallback to try kicking the closed
timestamp when it's behind. Since these are unexected, we also log them
and keep a metric.

Touches #35142

Release note (bug fix): Made `CHANGEFEED`s more resilient to a class of
bugs which manifest as stalls.