storage: detect intent timestamp violations when resolving intents
This commit adds detection for violations of the condition governing how
committed intent timestamps can related to the clock of their Range's
leaseholder. In order for the transaction observed timestamp optimization to
work without allowing stale reads in clusters with loosly synchronized clocks,
transaction's must never be committed at timestamps above that in the clock of
any leaseholder that contains any of its intents. See the comment on
roachpb.Transaction.ObservedTimestamps for a detailed discussion on why that
invariant must hold.

We can partially detect violations of that condition here by comparing the
commit timestamp of the intent's transaction with the timestamp of the local
clock.

Release note: None