[PROTOTYPE] storage: move intent meta to separate lock-table keyspace
Prototype of issue 41720 (no # because I don't want link to issue yet).

The main focus of this prototype is a new `pkg/storage/concurrency`
package, which provides a concurrency manager structure that encapsulates
the details of concurrency control and contention handling for
serializable key-value transactions. Any reader of this prototype should
start at `concurrency_control.go` and move out from there.

This fails a lot of tests, but can be run with:
  `cockroach start --insecure --storage-engine=pebble`

From my breif testing and benchmarking, I'm not seeing a noticable
regression in performance on non-transactional or transactional
workloads.