[DNM] kv: add trace events to lock wait-queue state transitions
This change adds trace events to each state transition observed by the
lockTableWaiter while waiting on a collection of lockWaitQueues. This
provides increased observability into the behavior of lock wait-queues
on contended keys, including information about queue depth.

We'll probably want to polish this off and merge it at some point, but for now,
it's being used for testing.