storage: ignore non-intent locks in intentInterleavingIter
This commit updates the intentInterleavingIter to ignore locks
in the lock table keyspace with strengths other than lock.Intent
(i.e. shared and exclusive locks). Future versions of the
iterator may expose information to users about whether any
non-intent locks were observed and, if so, which keys they were
found on. For now, no such information is exposed.

This change is needed for replicated locks. Non-locking reads
will not interact with these locks at all, so we want the
intentInterleavingIter to ignore them. Meanwhile, locking reads
and writes will use a different abstraction to peek into the
lock table and look for conflicts.

Release note: None