[DNM] storage: add dynamic iter/seek adjustment to intentInterleavingIter
This commit configures the intentInterleavingIter to dynamically adjust
the number of iterations that it will try across the locks on single
user key while searching for intents in the lock table before it does a
SeekGE/SeekLT to skip to the locks on the next/previous user key.

This places a bound on the amount of work that non-locking iteration
will perform for keys with many (shared) locks. As a result, scan speed
will no longer be proportional to the total number of locks on the keys
that it scans over. Scan speed will still be proportional to the number
of keys that have locks, but we expect that keys with locks will also
have values, so this per-locked-key cost is acceptable.

Don't merge, this needs more testing.