roachtest: run pgjdbc tests under correct user/database
The pgjdbc tests were running under the "test_admin" user in the
"defaultdb" database. This was tripping up some tests. We now run the
tests under the "test" user in the "test" database according to the
requirements of the pgjdbc test suite, which are outlined in
https://jdbc.postgresql.org/development/#test-suite.

Epic: None
Release note: None